Achieve +60% test coverage
