--- __tests__/sum.test.js
@@ -1,7 +1,0 @@
-function sum (a ,b) {
-  return a + b;
-}
-
-test('adds 1 + 2 to equal 3', () => {
-  expect(sum(1, 2)).toBe(3);
-});
