+++ jestsetup.js
@@ -1,0 +1,6 @@
+import enzyme from 'enzyme';
+import Adapter from 'enzyme-adapter-react-16';
+
+enzyme.configure({
+  adapter: new Adapter()
+});
